+++ src/fsharp/CompileOps.fs
@@ -4260,7 +4260,7 @@ type TcImports(tcConfigP:TcConfigProvider, initialResolutions:TcAssemblyResoluti
 
         if tcImports.IsAlreadyRegistered ilShortAssemName then 
             let dllinfo = tcImports.FindDllInfo(m,ilShortAssemName)
-            let phase2() = [tcImports.FindCcuInfo(m,ilShortAssemName,lookupOnly=false)]
+            let phase2() = [tcImports.FindCcuInfo(m,ilShortAssemName,lookupOnly=true)] 
             dllinfo,phase2
         else 
             let dllinfo = {RawMetadata=assemblyData 
@@ -4373,9 +4373,16 @@ type TcImports(tcConfigP:TcConfigProvider, initialResolutions:TcAssemblyResoluti
             | None ->      
                                   
                 if tcConfigP.Get().useMonoResolution then
-                    let resolved = [tcConfig.ResolveLibWithDirectories CcuLoadFailureAction.RaiseError assemblyReference |> Option.get]
-                    resolutions <- resolutions.AddResolutionResults resolved
-                    ResultD resolved
+                    let action = 
+                        match mode with 
+                        | ResolveAssemblyReferenceMode.ReportErrors -> CcuLoadFailureAction.RaiseError
+                        | ResolveAssemblyReferenceMode.Speculative -> CcuLoadFailureAction.ReturnNone
+                    match tcConfig.ResolveLibWithDirectories action assemblyReference with 
+                    | Some resolved -> 
+                        resolutions <- resolutions.AddResolutionResults [resolved]
+                        ResultD [resolved]
+                    | None -> 
+                        ErrorD(AssemblyNotResolved(assemblyReference.Text,assemblyReference.Range))
                 else 
                     // This is a previously unencounterd assembly. Resolve it and add it to the list.
                     // But don't cache resolution failures because the assembly may appear on the disk later.
@@ -5252,6 +5259,7 @@ let TypeCheckClosedInputSet (checkForErrors, tcConfig, tcImports, tcGlobals, pre
     let (tcEnvAtEndOfLastFile, topAttrs, mimpls),tcState = TypeCheckMultipleInputs (checkForErrors, tcConfig, tcImports, tcGlobals, prefixPathOpt, tcState, inputs)
     let tcState,tassembly = TypeCheckClosedInputSetFinish (mimpls, tcState)
     tcState, topAttrs, tassembly, tcEnvAtEndOfLastFile
+