+++ src/fsharp/vs/IncrementalBuild.fs
@@ -1059,7 +1059,7 @@ type FrameworkImportsCache(keepStrongly) =
 
 
 /// An error logger that capture errors, filtering them according to warning levels etc.
-type CompilationErrorLogger (debugName:string, tcConfig:TcConfig) = 
+type internal CompilationErrorLogger (debugName:string, tcConfig:TcConfig) = 
     inherit ErrorLogger("CompilationErrorLogger("+debugName+")")
             
     let warningsSeenInScope = new ResizeArray<_>()
@@ -1167,7 +1167,7 @@ type IncrementalBuilder(frameworkTcImportsCache: FrameworkImportsCache, tcConfig
     // This operation is done when constructing the builder itself, rather than as an incremental task. 
     let nonFrameworkAssemblyInputs = 
         // Note we are not calling errorLogger.GetErrors() anywhere for this task. 
-        // REVIEW: Consider if this is ok. I believe so, because this is a background build and we aren'T currently reporting errors from the background build. 
+        // This is ok because not much can actually go wrong here.
         let errorLogger = CompilationErrorLogger("nonFrameworkAssemblyInputs", tcConfig)
         // Return the disposable object that cleans up
         use _holder = new CompilationGlobalsScope(errorLogger,BuildPhase.Parameter, projectDirectory) 
@@ -1182,7 +1182,6 @@ type IncrementalBuilder(frameworkTcImportsCache: FrameworkImportsCache, tcConfig
                         DateTime.Now                               
                 with e -> 
                     // Note we are not calling errorLogger.GetErrors() anywhere for this task. This warning will not be reported...
-                    // REVIEW: Consider if this is ok. I believe so, because this is a background build and we aren't currently reporting errors from the background build. 
                     errorLogger.Warning(e)
                     DateTime.Now                               
             yield (Choice1Of2 r.resolvedPath,originalTimeStamp)  
@@ -1704,7 +1703,7 @@ type IncrementalBuilder(frameworkTcImportsCache: FrameworkImportsCache, tcConfig
                         else MSBuildResolver.CompileTimeLike
                 
                 tcConfigB.conditionalCompilationDefines <- 
-                    let define = if useScriptResolutionRules then "INTERACTIVE" else "COMPILED"
+                    let define = if useScriptResolutionRules then "INTERACTIVE" else "MPILED"
                     define::tcConfigB.conditionalCompilationDefines
 
                 tcConfigB.projectReferences <- projectReferences