+++ src/fsharp/vs/Symbols.fs
@@ -656,7 +656,7 @@ and FSharpField(cenv, d: FSharpFieldData)  =
         if isUnresolved() then false else 
         match d.TryRecdField with 
         | Choice1Of2 r -> r.IsMutable
-        | Choice2Of2 f -> not f.IsInitOnly
+        | Choice2Of2 f -> not f.IsInitOnly && f.LiteralValue.IsNone
 
     member __.IsLiteral = 
         if isUnresolved() then false else 
@@ -1403,6 +1403,15 @@ and FSharpMemberOrFunctionOrValue(cenv, d:FSharpMemberOrValData, item) =
         | P p -> not p.IsStatic
         | M m -> m.IsInstance
         | V v -> v.IsInstanceMember
+
+    member v.IsInstanceMemberInCompiledCode = 
+        if isUnresolved() then false else 
+        v.IsInstanceMember &&
+        match d with 
+        | E e -> match e.ArbitraryValRef with Some vref -> ValRefIsCompiledAsInstanceMember cenv.g vref | None -> true
+        | P p -> match p.ArbitraryValRef with Some vref -> ValRefIsCompiledAsInstanceMember cenv.g vref | None -> true
+        | M m -> match m.ArbitraryValRef with Some vref -> ValRefIsCompiledAsInstanceMember cenv.g vref | None -> true
+        | V vref -> ValRefIsCompiledAsInstanceMember cenv.g vref 
 
     member __.IsExtensionMember = 
         if isUnresolved() then false else