+++ RELEASE_NOTES.md
@@ -1,3 +1,16 @@
+#### 2.0.0.1-beta 
+* Fix 452 - FSharpField.IsMutable = true for BCL enum cases
+* Fix 452 - FSharpField.IsMutable = true for BCL enum cases
+* Fix 414 - Add IsInstanceMemberInCompiledCode
+
+#### 2.0.0.0-beta 
+* Feature #470, #478, #479 - Move ProjectCracker to separate nuget package and DLL, used ProjectCrackerTool.exe to run 
+* Feature #463 - Expose slot signatures of members in object expressions
+* Feature #469, #475 - Add EvalExpressionNonThrowing, EvalInteractionNonThrowing, EvalScriptNonThrowing 
+* Fix #456 - FCS makes calls to kernel32.dll when running on OSX
+* Fix #473 - stack overflow in resolution logic
+* Fix #460 - Failure getting expression for a provided method call
+
 #### 1.4.2.3 - 
 * Fix bug in loop optimization, apply https://github.com/Microsoft/visualfsharp/pull/756/